Fix split package issue in Kubernetes Service Binding extension
Fixes: #44732
