+++ extensions/kubernetes-service-binding/deployment/src/main/java/io/quarkus/kubernetes/service/binding/deployment/AddServiceBindingResourceDecorator.java
@@ -1,5 +1,5 @@
 
-package io.quarkus.kubernetes.service.binding.buildtime;
+package io.quarkus.kubernetes.service.binding.deployment;
 
 import java.util.Optional;
