+++ extensions/kubernetes-service-binding/deployment/src/main/java/io/quarkus/kubernetes/service/binding/deployment/KubernetesServiceBindingProcessor.java
@@ -1,8 +1,9 @@
-package io.quarkus.kubernetes.service.binding.runtime;
+package io.quarkus.kubernetes.service.binding.deployment;
 
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.RunTimeConfigBuilderBuildItem;
+import io.quarkus.kubernetes.service.binding.runtime.KubernetesServiceBindingConfigSourceFactoryBuilder;
 
 public class KubernetesServiceBindingProcessor {
     @BuildStep
