+++ extensions/kubernetes-service-binding/deployment/src/main/java/io/quarkus/kubernetes/service/binding/deployment/ServiceBindingProcessor.java
@@ -1,5 +1,5 @@
 
-package io.quarkus.kubernetes.service.binding.buildtime;
+package io.quarkus.kubernetes.service.binding.deployment;
 
 import java.util.ArrayList;
 import java.util.HashMap;
